Accept hip stream legacy (#689)
* Support getting device ID from stream hipStreamLegacy (#687)

rocprim::get_device_from_stream must retreive the device ID
in different ways depending on whether a user-created stream
or a built-in stream type is being used.

This change modifies the check for built-in stream types so that
it includes hipStreamLegacy. This prevents the function from
calling hipGetStreamDeviceId(hipStreamLegacy), which currently
causes a segmentation fault.

* Add hipStreamLegacy HIP version check

The hipStreamLegacy stream handle was added in HIP 6.1.0.
Add a check to ensure that we don't reference it if rocPRIM
is being built using an older version of HIP.

* Add get_device_id_from_stream test for hipStreamLegacy

Add a test case to ensure that passing hipStreamLegacy to
rocprim::get_device_id_from_stream works correctly.

Updated the changelog with a note to reflect the fact that
hipStreamLegacy can now be used without issue.

* Update CHANGELOG.md

Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>

---------

Co-authored-by: Jeffrey Novotny <jnovotny@amd.com>